requirements.txt
openai
python-dotenv
pytest
tiktoken